Add option for keeping local cache of recent data.  (#585)
This PR adds local cache / hot tier for Parseable.

This option can be enabled by setting following env vars
P_CACHE_DIR - Local Path for file cache
P_CACHE_SIZE - Size for cache in human readable size ( mb/mib/gib/gb )

When these flags are set, sync flow will move the parquet from staging into
the specified cache directory instead of deleting it. Any LRU cached entry is
deleted to satisfy the cache constraint upon insertion. LocalCacheManager is
responsible for updating and persisting this data structure.

+++ server/src/storage/object_storage.rs
@@ -24,6 +24,7 @@ use super::{
 use crate::{
     alerts::Alerts,
     catalog::{self, manifest::Manifest, snapshot::Snapshot},
+    localcache::LocalCacheManager,
     metadata::STREAM_INFO,
     metrics::{storage::StorageMetrics, STORAGE_SIZE},
     option::CONFIG,
@@ -35,6 +36,7 @@ use arrow_schema::Schema;
 use async_trait::async_trait;
 use bytes::Bytes;
 use datafusion::{datasource::listing::ListingTableUrl, execution::runtime_env::RuntimeConfig};
+use itertools::Itertools;
 use relative_path::RelativePath;
 use relative_path::RelativePathBuf;
 use serde_json::Value;
@@ -196,6 +198,15 @@ pub trait ObjectStorage: Sync + 'static {
         Ok(serde_json::from_slice(&stream_metadata).expect("parseable config is valid json"))
     }
 
+    async fn put_stream_manifest(
+        &self,
+        stream_name: &str,
+        manifest: &ObjectStoreFormat,
+    ) -> Result<(), ObjectStorageError> {
+        let path = stream_json_path(stream_name);
+        self.put_object(&path, to_bytes(manifest)).await
+    }
+
     async fn get_stats(&self, stream_name: &str) -> Result<Stats, ObjectStorageError> {
         let stream_metadata = self.get_object(&stream_json_path(stream_name)).await?;
         let stream_metadata: Value =
@@ -218,7 +229,6 @@ pub trait ObjectStorage: Sync + 'static {
             .expect("is object")
             .get("retention")
             .cloned();
-
         if let Some(retention) = retention {
             Ok(serde_json::from_value(retention).unwrap())
         } else {
@@ -306,10 +316,15 @@ pub trait ObjectStorage: Sync + 'static {
         }
 
         let streams = STREAM_INFO.list_streams();
-
         let mut stream_stats = HashMap::new();
 
+        let cache_manager = LocalCacheManager::global();
+        let mut cache_updates: HashMap<&String, Vec<_>> = HashMap::new();
+
         for stream in &streams {
+            let cache_enabled = STREAM_INFO
+                .cache_enabled(stream)
+                .map_err(|err| ObjectStorageError::UnhandledError(Box::new(err)))?;
             let dir = StorageDir::new(stream);
             let schema = convert_disk_files_to_parquet(stream, &dir)
                 .map_err(|err| ObjectStorageError::UnhandledError(Box::new(err)))?;
@@ -327,27 +342,30 @@ pub trait ObjectStorage: Sync + 'static {
                     .or_insert_with(|| compressed_size);
             });
 
-            for file in &parquet_files {
+            for file in parquet_files {
                 let filename = file
                     .file_name()
                     .expect("only parquet files are returned by iterator")
                     .to_str()
                     .expect("filename is valid string");
                 let file_suffix = str::replacen(filename, ".", "/", 3);
-                let objectstore_path = format!("{stream}/{file_suffix}");
-                let manifest = catalog::create_from_parquet_file(
-                    self.absolute_url(RelativePath::from_path(&objectstore_path).unwrap())
-                        .to_string(),
-                    file,
-                )
-                .unwrap();
-                self.upload_file(&objectstore_path, file).await?;
+                let stream_relative_path = format!("{stream}/{file_suffix}");
+                self.upload_file(&stream_relative_path, &file).await?;
+                let absolute_path = self
+                    .absolute_url(RelativePath::from_path(&stream_relative_path).unwrap())
+                    .to_string();
                 let store = CONFIG.storage().get_object_store();
+                let manifest =
+                    catalog::create_from_parquet_file(absolute_path.clone(), &file).unwrap();
                 catalog::update_snapshot(store, stream, manifest).await?;
-            }
-
-            for file in parquet_files {
-                let _ = fs::remove_file(file);
+                if cache_enabled && cache_manager.is_some() {
+                    cache_updates
+                        .entry(stream)
+                        .or_default()
+                        .push((absolute_path, file))
+                } else {
+                    let _ = fs::remove_file(file);
+                }
             }
         }
 
@@ -361,6 +379,24 @@ pub trait ObjectStorage: Sync + 'static {
                     log::warn!("Error updating stats to objectstore due to error [{}]", e);
                 }
             }
+        }
+
+        if let Some(manager) = cache_manager {
+            let cache_updates = cache_updates
+                .into_iter()
+                .map(|(key, value)| (key.to_owned(), value))
+                .collect_vec();
+
+            tokio::spawn(async move {
+                for (stream, files) in cache_updates {
+                    for (storage_path, file) in files {
+                        manager
+                            .move_to_cache(&stream, storage_path, file.to_owned())
+                            .await
+                            .unwrap()
+                    }
+                }
+            });
         }
 
         Ok(())
